Implements Longest Word
Simple program, so only one commit is necessary here.

Uses a Trie to implement this algorithm.

Uses a python dict (hash table) to perform the final lookup.
This should save a lot of time and computation.
Since we are only looking for the largest compund word, we
can keep a hash that keys off word length with a list of words
of that length as it's values.
We just reverse the list of keys for that hash, and stop as soon
as we have our first match.